+++ lib/features/journal/repository/journal_repository.dart
@@ -5,7 +5,6 @@ import 'package:lotti/classes/entry_text.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/conversions.dart';
 import 'package:lotti/database/database.dart';
-import 'package:lotti/features/ai/services/task_summary_refresh_service.dart';
 import 'package:lotti/features/sync/model/sync_message.dart';
 import 'package:lotti/features/sync/outbox/outbox_service.dart';
 import 'package:lotti/get_it.dart';
@@ -20,20 +19,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 part 'journal_repository.g.dart';
 
 class JournalRepository {
-  JournalRepository(this._ref);
-
-  final Ref _ref;
-
-  /// Triggers a task summary refresh for all tasks linked to the given checklist
-  Future<void> _triggerTaskSummaryRefreshForChecklist(
-      String checklistId) async {
-    await _ref
-        .read(taskSummaryRefreshServiceProvider)
-        .triggerTaskSummaryRefreshForChecklist(
-          checklistId: checklistId,
-          callingDomain: 'JournalRepository',
-        );
-  }
+  JournalRepository();
 
   /// Clears coverArtId from any tasks that reference the deleted image
   Future<void> _clearCoverArtReferences(
@@ -105,13 +91,6 @@ class JournalRepository {
 
       if (journalEntity == null) {
         return false;
-      }
-
-      // If deleting a checklist item, trigger task summary refresh
-      if (journalEntity is ChecklistItem) {
-        for (final checklistId in journalEntity.data.linkedChecklists) {
-          await _triggerTaskSummaryRefreshForChecklist(checklistId);
-        }
       }
 
       // If deleting an image that is used as cover art, clear the coverArtId
@@ -379,4 +358,4 @@ class JournalRepository {
 }
 
 @riverpod
-JournalRepository journalRepository(Ref ref) => JournalRepository(ref);
+JournalRepository journalRepository(Ref ref) => JournalRepository();