+++ lib/features/tasks/repository/checklist_repository.g.dart
@@ -51,4 +51,4 @@ final class ChecklistRepositoryProvider extends $FunctionalProvider<
 }
 
 String _$checklistRepositoryHash() =>
-    r'ddd5a084b352361abb5a29fbe1ab58099dbc9585';
+    r'01da903b25b6648bc4f1b17cbb3915e03c09c5bd';